Merging the cleaned up version into master (#1)
* add exercises for training session finished

Could use a search bar. The training list is hardcoded. there are 10 or so. It needs to read from a file instead or use a database.

* Clean up

Removed unnecessary files and imports

* reformating myHomePag to add_exercises_state instead

reformating to external class instead of being located in main

* Changed myhomepage to homepag and refactored _myhomepagestate to AddExercisesState

* adding firebase core and database

tried changing several gradle files and doing fixes. It does not work currently as it cannot establish an connection.  https://firebase.google.com/docs/flutter/setup?platform=android#add-flutterfire-plugins

* removed imports